+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Changed
+
+- cli: Set environment variables from cli to spawned process instead of layer when using `mirrord exec`.
+- cli: use miette for nicer errors
+- cli: some ext exec preparations, nothing user facing yet.
+
 ## 3.13.5
 
 ### Changed